refactor: organise imports and exports in layer-2 module
