+++ packages/shared/src/lib/core/activity/utils/helper/getSubjectFromActivity.ts
@@ -1,7 +1,7 @@
 import { ActivityType } from '@core/activity/enums'
 import { Activity } from '@core/activity/types'
 import { localize } from '@core/i18n'
-import { getLayer2NetworkFromAddress } from '@core/layer-2/utils'
+import { getLayer2NetworkFromAddress } from '@core/layer-2/actions'
 import { truncateString } from '@core/utils'
 import { SubjectType } from '@core/wallet/enums'
 import type { Subject } from '@core/wallet/types'
