+++ packages/shared/src/lib/core/layer-2/actions/getLayer2NetworkFromAddress.ts
