+++ packages/shared/src/lib/core/layer-2/actions/index.ts
@@ -1,7 +1,14 @@
+export * from './buildEvmTransactionData'
+export * from './buildUnwrapAssetParameters'
+export * from './buildUnwrapAssetTargetAddress'
 export * from './estimateGasForLayer1ToLayer2Transaction'
 export * from './fetchSelectedAccountLayer2Balance'
 export * from './generateAndStoreEvmAddressForAccount'
+export * from './getDestinationNetworkFromAddress'
 export * from './getGasFeesForLayer1ToLayer2Transaction'
 export * from './getGasPriceInWei'
+export * from './getIscpTransferSmartContractData'
 export * from './getLayer2MetadataForTransfer'
+export * from './getLayer2NetworkFromAddress'
 export * from './pollLayer2Tokens'
+export * from './signEvmTransactionWithStronghold'
