+++ packages/shared/src/lib/core/layer-2/actions/signEvmTransactionWithStronghold.ts
