+++ packages/shared/src/lib/core/layer-2/helpers/evmAddressToAgentId.ts
@@ -1,6 +1,6 @@
 import { Converter } from '@iota/util.js'
 
-export function evmAddressToAgentID(evmStoreAccount: string): Uint8Array {
+export function evmAddressToAgentId(evmStoreAccount: string): Uint8Array {
     // This function constructs an AgentID that is required to be used with contracts
     // Wasp understands different AgentID types and each AgentID needs to provide a certain ID that describes it's address type.
     // In the case of EVM addresses it's ID 3.
