+++ packages/shared/src/lib/core/layer-2/helpers/getEvmTransactionValueFromAmount.ts
