+++ packages/shared/src/lib/core/layer-2/helpers/index.ts
@@ -1,3 +1,10 @@
+export * from './buildAssetAllowance'
+export * from './calculateGasFeeInGlow'
 export * from './encodeAddress'
 export * from './encodeAssetAllowance'
 export * from './encodeSmartContractParameters'
+export * from './evmAddressToAgentId'
+export * from './getAgentBalanceParameters'
+export * from './getEvmTransactionValueFromAmount'
+export * from './getSmartContractHexName'
+export * from './specialNativeTokenAmountEncoding'
