+++ packages/shared/src/lib/core/layer-2/helpers/specialNativeTokenAmountEncoding.ts
