+++ packages/shared/src/lib/core/layer-2/index.ts
@@ -1,5 +1,4 @@
 export * from './abis'
-export * from './actions'
 export * from './constants'
 export * from './classes'
 export * from './constants'
