+++ packages/shared/src/lib/core/layer-2/utils/calculateEstimatedGasFeeFromTransactionData.ts
@@ -1,6 +1,6 @@
 import { BigIntLike } from '@ethereumjs/util'
 import { EvmTransactionData } from '../types'
-import { calculateGasFeeInGlow } from './calculateGasFeeInGlow'
+import { calculateGasFeeInGlow } from '../helpers'
 
 export function calculateEstimatedGasFeeFromTransactionData(
     transactionData: EvmTransactionData
