+++ packages/shared/src/lib/core/layer-2/utils/index.ts
@@ -1,23 +1,9 @@
-export * from './buildAssetAllowance'
-export * from './buildEvmTransactionData'
-export * from './buildUnwrapAssetParameters'
-export * from './buildUnwrapAssetTargetAddress'
 export * from './calculateEstimatedGasFeeFromTransactionData'
-export * from './calculateGasFeeInGlow'
 export * from './calculateMaxGasFeeFromTransactionData'
-export * from './evmAddressToAgentID'
 export * from './getAbiForContractType'
-export * from './getAgentBalanceParameters'
 export * from './getAmountFromEvmTransactionValue'
-export * from './getDestinationNetworkFromAddress'
 export * from './getErc20TokenMetadata'
 export * from './getErc20TransferSmartContractData'
-export * from './getEvmTransactionValueFromAmount'
-export * from './getIscpTransferSmartContractData'
-export * from './getLayer2NetworkFromAddress'
-export * from './getSmartContractHexName'
 export * from './parseLayer2Metadata'
 export * from './parseLayer2MetadataForTransfer'
 export * from './prepareEvmTransaction'
-export * from './signEvmTransactionWithStronghold'
-export * from './specialNativeTokenAmountEncoding'
