+++ packages/shared/src/lib/core/wallet/actions/send/createEvmTransactionFromSendFlowParameters.ts
@@ -1,15 +1,15 @@
 import { IAccountState } from '@core/account'
-import { AssetType } from '@core/layer-2'
-import { ISC_MAGIC_CONTRACT_ADDRESS } from '@core/layer-2/constants'
-import { EvmTransactionData, TransferredAsset } from '@core/layer-2/types'
 import {
-    buildEvmTransactionData,
+    AssetType,
+    EvmTransactionData,
+    ISC_MAGIC_CONTRACT_ADDRESS,
+    TransferredAsset,
     getErc20TransferSmartContractData,
-    getIscpTransferSmartContractData,
-} from '@core/layer-2/utils'
+} from '@core/layer-2'
+import { buildEvmTransactionData, getIscpTransferSmartContractData } from '@core/layer-2/actions'
 import { IChain } from '@core/network/interfaces'
-import { IToken } from '@core/token/interfaces'
 import { TokenStandard } from '@core/token/enums'
+import { IToken } from '@core/token/interfaces'
 import { SendFlowType } from '@core/wallet/stores'
 import { SendFlowParameters } from '@core/wallet/types'
