+++ packages/shared/src/lib/core/wallet/actions/send/signAndSendEvmTransaction.ts
@@ -2,7 +2,7 @@ import { IAccountState } from '@core/account'
 import { updateSelectedAccount } from '@core/account/stores'
 import { handleError } from '@core/error/handlers'
 import { EvmTransactionData } from '@core/layer-2/types'
-import { signEvmTransactionWithStronghold } from '@core/layer-2/utils'
+import { signEvmTransactionWithStronghold } from '@core/layer-2/actions'
 import { Ledger } from '@core/ledger/classes'
 import { EvmChainId } from '@core/network/enums'
 import { isActiveLedgerProfile, isSoftwareProfile } from '@core/profile/stores'
